Create expression for total num of stores in loop.
If all stores in a loop are unconditional and the loop backedge
taken count is invariant, then we can use SCEV to create an
expression for the total number of stores in the loop. Then that
expression can be used to increment the stValTag in the main loop,
instead of waiting for the PE loop to communicate back its new tag.